optimize SIMD performance ~10%
