+++ src/array/primitive_array.rs
@@ -1,8 +1,13 @@
 use super::{Array, ArrayBuilder, ArrayValidExt};
 use crate::types::NativeType;
-use bitvec::prelude::{BitSlice, BitVec, Lsb0};
+use bitvec::vec::BitVec;
 use serde::{Deserialize, Serialize};
 use std::iter::FromIterator;
+
+#[cfg(feature = "simd")]
+mod simd;
+#[cfg(feature = "simd")]
+pub use self::simd::*;
 
 /// A collection of primitive types, such as `i32`, `f32`.
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -83,120 +88,10 @@ impl<T: NativeType> ArrayBuilder for PrimitiveArrayBuilder<T> {
     }
 }
 
-impl<T: NativeType> PrimitiveArray<T> {
-    /// Returns a batch iterator for SIMD.
-    ///
-    /// Each item contains at most `N` elements.
-    pub fn batch_iter<const N: usize>(&self) -> BatchIter<'_, T, N> {
-        assert!(N <= std::mem::size_of::<usize>() * 8);
-        BatchIter {
-            array: self,
-            idx: 0,
-        }
-    }
-}
-
-/// An iterator over a batch elements of the array at a time.
-pub struct BatchIter<'a, T: NativeType, const N: usize> {
-    array: &'a PrimitiveArray<T>,
-    idx: usize,
-}
-
-/// A batch elements generated by `BatchIter`.
-#[derive(Debug, PartialEq, Eq)]
-pub struct BatchItem<T: NativeType, const N: usize> {
-    /// The valid (non-NULL) bitmap.
-    pub valid: usize,
-    /// The elements.
-    pub data: [T; N],
-    /// The length of the batch.
-    pub len: usize,
-}
-
-impl<T: NativeType, const N: usize> Iterator for BatchIter<'_, T, N> {
-    type Item = BatchItem<T, N>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.idx >= self.array.len() {
-            return None;
-        }
-        let len = (self.array.len() - self.idx).min(N);
-        let range = self.idx..self.idx + len;
-
-        assert_eq!(N % 8, 0, "batch size must be a multiple of 8");
-        let mut valid = [0u8; std::mem::size_of::<usize>()];
-        let bytes = (len + 7) / 8;
-        valid[..bytes].copy_from_slice(unsafe {
-            std::slice::from_raw_parts(
-                (self.array.valid.as_raw_ptr() as *const u8).add(self.idx / 8),
-                bytes,
-            )
-        });
-        let valid = usize::from_le_bytes(valid);
-
-        let data: [T; N] = if len == N {
-            self.array.data[range].try_into().unwrap()
-        } else {
-            let mut data = [T::ZERO; N];
-            data[..len].copy_from_slice(&self.array.data[range]);
-            data
-        };
-
-        self.idx += N;
-        Some(BatchItem { valid, data, len })
-    }
-}
-
-impl<T: NativeType, const N: usize> FromIterator<BatchItem<T, N>> for PrimitiveArray<T> {
-    fn from_iter<I: IntoIterator<Item = BatchItem<T, N>>>(iter: I) -> Self {
-        let iter = iter.into_iter();
-        let mut builder = PrimitiveArrayBuilder::new(iter.size_hint().0 * N);
-        for e in iter {
-            builder
-                .valid
-                .extend_from_bitslice(&BitSlice::<Lsb0, usize>::from_element(&e.valid)[..e.len]);
-            builder.data.extend_from_slice(&e.data[..e.len]);
-        }
-        builder.finish()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use num_traits::cast::FromPrimitive;
-
-    #[test]
-    fn batch_iter() {
-        let a = (0..12)
-            .map(|i| if i % 2 == 0 { Some(i) } else { None })
-            .collect::<PrimitiveArray<u32>>();
-        let mut iter = a.batch_iter::<8>();
-        assert_eq!(
-            iter.next(),
-            Some(BatchItem {
-                valid: 0b_0101_0101,
-                data: [0, 0, 2, 0, 4, 0, 6, 0],
-                len: 8
-            })
-        );
-        assert_eq!(
-            iter.next(),
-            Some(BatchItem {
-                valid: 0b_0000_0101,
-                data: [8, 0, 10, 0, 0, 0, 0, 0],
-                len: 4
-            })
-        );
-        assert_eq!(iter.next(), None);
-    }
-
-    #[test]
-    fn batch_iter_collect() {
-        let a = (0..12).collect::<PrimitiveArray<u32>>();
-        let a1 = a.batch_iter::<8>().collect::<PrimitiveArray<u32>>();
-        assert_eq!(a1, a);
-    }
 
     fn test_builder<T: FromPrimitive + NativeType>() {
         let iter = (0..1000).map(|x| if x % 2 == 0 { None } else { T::from_usize(x) });
